Fix UBSan errors in wtf_unittests
Most of this is the C language bug where memcpy and friends cannot take
nullptr, even if length is zero. Elsewhere, I've been replacing them
with <algorithm> functions, which don't have this bug. But <algorithm>
is aware of the underlying type, while wtf/vector.h seems to have some
trait for types to opt out of the language-level constructor machinery.

I suspect all this (slightly unsafe) machinery could be made unnecessary
if we helped the optimizer out a bit more. E.g. if all the
memset/memcpy-compatible types made their trivial constructors and
trivial copy constructors visible to the header (counter to Chromium's
Clang plugin), the compiler would know that initializing them is trivial
and then emit the right code. But that's a much larger lift so, for now,
just skip memcpy if the span is empty.

And then atomic_operations_test.cc needs to be switched from
type-punning via pointer casts to type-punning via memcpy (here, memcpy
is the function we want) because the pointer casts are UB via strict
aliasing (disabled in Chromium) and additionally UB if the pointer is
not correctly aligned.

Bug: 1394755
Change-Id: I97f706279668521f9df7728a9129856f5b879a0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4973517
Auto-Submit: David Benjamin <davidben@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214607}