input_page: Show language packs error message on IMEs
See go/language-pack-ux for mocks. Note that we do not have a "Try
again" button, as that is pending discussion.

The message is currently marked as non-translateable as the message is
misleading for the current state of language packs for input methods
(only handwriting).

Bug: b:298128915
Change-Id: I1a4e7cc5a994914faaa7778f751268dbba8a31dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4963047
Commit-Queue: Michael Cui <mlcui@google.com>
Reviewed-by: Claudio M <claudiomagni@chromium.org>
Reviewed-by: David Padlipsky <dpad@google.com>
Cr-Commit-Position: refs/heads/main@{#1214658}