[ios] Move ShortClickInteractionTimestamp key out of NSUserDefaults
This CL moves ShortClickInteractionTimestamp out of NSUserDefaults.

The key has been moved in the PrefService. Note that there are two
types of PrefServices, in this case the code was moved in the
PrefService coming from BrowserState which is per account and is
syncable between devices.
Note that this kind of PrefService behaves differently from
NSUserDefaults which is per device (and not per account).

Bug: 1486770
Change-Id: I4031a46861a308a4464865a395353b39b709f036
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4965920
Reviewed-by: Guillem Perez <guiperez@google.com>
Commit-Queue: Federica Germinario <fedegermi@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1214697}