Test that background tabs are not immediately considered inactive
This CL adds an EG test to prevent regressing against crbug/1490604.
When a tab was opened in the background but never displayed, it was
considered inactive, so on the next launch, it was moved to Inactive
Tabs.

Fixed: 1492488
Change-Id: I22b40d7f6caf2f3eafe1e4bdcc18fee5c2941ab0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4968842
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1214702}