iwa: Make tests for IWA install command more idiomatic (3/3)
This is the third step towards using the proper testing APIs provided by
the Web App system. In this CL, we mainly replace the use of a fake
`IsolatedWebAppResponseReaderFactory`, and we no longer construct the
command object by ourselves - instead the `WebAppCommandScheduler` is
now taking care of it. This also adds additional bundle-based tests for
error cases.

Fixed: b:288395295
Bug: b:288395295
Change-Id: I8b5300062ae654982303d5bff6b680f412a731da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4966707
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214713}