Deprecate vector<v8::Local>, part 6
According to V8's public API documentation, local handles (i.e.,
objects of type v8::Local<T>) "should never be allocated on the heap".
This disallows heap-allocated data structures containing instances of
v8::Local, like std::vector<v8::Local<v8::String>>. It is unfortunate
that the V8 API itself requires the usage of such data structures.

This CL continues the deprecation of such data structures. It replaces
such vectors with v8::LocalVector<T>, recently introduced in V8 API,
which behave exactly the same with the exception that the V8 isolate
must be passed upon construction.

Most of the changes in this CL are trivial and are related to the
renderer.

Bug: chromium:1454114
Change-Id: I26a42af7b4d2a2e9480d28b07088579e9fdded2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4922577
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214715}