Remove BAM::FillOrPreviewForm(Suggestion::BackendId)
The function retrieves the data model (profile or card) from the PDM
by the backend id and calls either FillOrPreviewProfileForm() or
FillOrPreviewCreditCardForm(). It only has a single non-test caller in
the external delegate.

This CL inlines the implementation there and migrates all callers to
either FillOrPreviewProfileForm() or FillOrPreviewCreditCardForm() (in
almost all cases, it's clear which one should be called).

This makes the external delegate slightly more complicated, but is
arguably the right change, since it simplifies the manager and our
filling API. (And it's what non-popup filling surfaces like TTF and
FastCheckout do).

The external delegate unit tests require some adaption: Currently, they
simply expect that FillOrPreviewForm() is called. These expectations are
changed to FillOrPreviewProfileForm() or FillOrPreviewCreditCardForm().
However, these functions are only called if the corresponding data model
is found in the PDM. Right now, the test setup doesn't take care of
that - and if fact doesn't even set backend ids properly.
To fix that, the tests are adapted by manually adding a profile (or
card) to the PDM and setting the backend ids of the suggestions
accordingly.

Bug: 1330108
Change-Id: I4c1110883adb3502efdad950e877d9d7cf99016d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4958776
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1214727}