Introduce CrdSessionObserver
In the remote commands code there are multiple reasons why we're
interested in certain CRD host events (host started/stopped and client
connected/disconnected).
Previously this was handled through a set of passed-in callbacks, but
that doesn't scale well if there are multiple reasons for wanting to
observe the same event, so I switched this to an observer based system.

Bug: b:266065062
Change-Id: I3e3779d956f28ffa980cebd03bf5d08d4254dad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4946413
Auto-Submit: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Ashutosh Singhal <macinashutosh@google.com>
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Artem Sumaneev <asumaneev@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214731}