Support kCreateNewPlusAddress autofill clicks on iOS
This is a v0 implementation, which just gets the plus address and
fills directly, rather than orchestrating the UI. It does allow
a simple test path for the AutofillDriverIOS::ApplyFieldAction,
which has not yet been implemented, but needs to be soon. This
CL aims to decouple that implementation from the bottom sheet UI.

Note that the ChromeAutofillClientIOS generally lacks unittest
coverage, and this feature isn't end-to-end testable yet. Earlgrey
tests will fill in the gap in the near future.

Low-Coverage-Reason:TESTS_IN_SEPARATE_CL

Bug: 1467623, b/295076363
Change-Id: I576bedb09a1ea7d8834729360aee9786e3ed06fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4969095
Commit-Queue: Matt Reichhoff <mreichhoff@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214762}