[mullet m2] Email 'fields' are also part of group filling.
For the regular mullet m2 case this makes no difference to the user.
This is however needed for group swapping, which will allow users to
change specific groups after a form was filled, the flow is:

1. User clicks on a field and fills the form.
2. User clicks on a name field and sees autofill suggestions.
3. Upon choosing a suggestion, only name fields are updated,
therefore group swapping.

To get this to work we also need to include email "fields" in the
granular filling groups list, since we will swap the targeting group.

Bug: 1459990
Change-Id: I9cd7981534fc63a3316866cce59bc3e406b74135
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4938611
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Bruno Braga <brunobraga@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1214765}