[UPMLocalSettings] Add the UPM readiness pref
PasswordsUseUPMLocalAndSeparateStores is a pref that signals if the user
has been migrated to UPM for local passwords (for both passwords and
settings).
This is a separate pref from kSettingsMigratedToUPMLocal, which is the
settings migration pref, used only by the settings service to decide
whether it still needs to migrate settings to GMS Core.

Bug: 1485511
Change-Id: I4a75275a35747a9b1f20571943df1f37a3704601
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4965481
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214773}