Fix operator== use in site_per_process_hit_test_browsertest.cc
Clang recently received a bug fix for how it finds operator==, causing
it to error on this code.

It seems like there was a mistake in this code, which is supposed to
compare the cursor's *type*, and the Clang bug made the compiler
silently accept it before.

Bug: 1495664
Change-Id: I5bd54ac2c7d5023af31d380a3b8403011ae6f14c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4974103
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Auto-Submit: Hans Wennborg <hans@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214798}