[dPWA] Implement desktop version of intent_picker delegate
This CL implements the desktop version of the intent_picker delegate, and hooks it up to the IntentPickerTabHelper to isolate behavior across CrOS and WML platforms. In doing so:
1. The intent_picker_helpers have been removed, and all behavior
moved to the delegates.
2. The intent_handling_metrics have been fully resolved of the
nogncheck.
3. Most of the remaining things were small changes and implementations
required to ensure that everything is hooked up fine.

Future CLs will include:
1. Tests for the delegate.
2. WebAppRegistrar tests.
3. Moving the desktop UX under the kDesktopPWAsLinkCapturing flag.
4. Other fixes required.

Bug: 1412856, b/304411994
Change-Id: I757ab7ebd1c3de0348dbb2fa308bc10b26882135
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4919875
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214881}