New field on chrome://policy/test for sign in restrictions policies
Add a new field to set the response from the profile separation
when getting them using UserCloudSigninRestrictionPolicyFetcher.

The chrome://policy/test page is the best place for this since the
page is used to test policies. This will only be available on Canary
after enabling the page via a flag.

This CL will allow local testing of the new profile separation
while waiting for the external endpoint to have them.

Bug: b:307738016
Change-Id: I6b332863b5532b267934b65ba3f1ac79fcc78437
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4973777
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Yann Dago <ydago@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214898}