[Base] add runNullSafe to Callback.java
Add a syntactic sugar static method to Callback that runs a possibly
null callback with a payload.

This can be used to simplify
if (callback != null) {
    callback.onResult(object);
}
to
Callback.runNullSafe(callback, object);

This is useful if an optional callback is provided to an interface.

See the ported examples in HubLayout.java for an example where this is
useful.

Change-Id: I5d525023a975029193a3f9532470c04ffb3a0acf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4967874
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214925}