Backlight OCR: Starts the work of implementing the "scroll" actions
Makes AXMediaAppHandler into an AXActionHandlerBase so that
accessibility actions would be sent to it for processing.
Specifically, in a future patch we aim to implement
the "scroll" actions, but given the size of the resulting work
it was considered appropriate to break this into two patches.

Please note that other actions cannot or should not be supported - see
design doc.

Design doc:
https://docs.google.com/document/d/1k8OH4XGjZrLVA705Xfy6vzSwGL9DkfRt9DG5s66lRjQ/edit?usp=sharing&resourcekey=0-K_9Y1EMhNoo8Ez9RPbQNMA

R=aleventhal@chromium.org

AX-Relnotes: n/a.
bug: b/289012145
Change-Id: I9d4095b19e51a3054fa1ef3bfde42dc4c8e75aad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4956596
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214927}