Use reinterpretColorSpace for tonemapped SkImages
When we draw texture draw quad, sometimes we need to use SkFilters to
do manual color covenrsion (e.g to tonemap HDR), in this case we pretend
that source SkImage has the same color space as destination to disable
the automatic color conversion on skia side and do it manually.

Before this CL we would pass override_color_space to ImageContextImpl
and cache it inside, which doesn't really work if the destination
color space changes. This CL switches to use reinterpretColorSpace()
instead.

Bug: 1494613
Change-Id: I5b6a60776d7f0bc9d1849df3296c721741e36075
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4973737
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214928}