Added flag for adding logs to the Autofill Rater Extension Report
This flag will enable chrome to listen for native messages from the
Autofill Rater Extension.

The messages will be of the type: "Report started", "Report ended".
When the report is started, Autofill and Password Manager Internals logs
will be recorded on a file on the disk. When it ends, the recording
stops.

b/304690377

Change-Id: Id1243c77d573d5312821a3dd304c8bc3224ca285
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4974451
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Theo Cristea <theocristea@google.com>
Cr-Commit-Position: refs/heads/main@{#1214935}