Plug the resource exhausted callback into Headless
On Windows, it is possible that the system is in a state where resources
are no longer available and we can't recover from that state. An
example are the ATOM for the window classes.

This CL is a follow up on the similar cleanup for the Chrome browser.
  https://crrev.com/c/4778839

R=grt@chromium.org

Bug: 1492444
Change-Id: I73d597d7660e94b6eb7b17ceb502281109e50bef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4966861
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214948}