Create RootGeometry lazily in IntersectionGeometry
When we call IntersectionObservation::ComputeIntersection in a loop
for intersection observations of the same intersection observer, we
use the same RootGeometry for all invocations.

Now let the caller create optional<RootGeometry> and pass it to
ComputeIntersection, and IntersectionGeometry emplace it only when
needed.

This simplifies the code by combining the previous two versions of
IntersectionObservation::ComputeIntersection() and
IntersectionGeometry constructors into one, respectively.

This performance impact will be probably trivial for now. This
prepares for IntersectionOptimization to track minimum scroll delta
per IntersectionObservation, with which ComputeIntersection will
have more no-op invocations and layily creating RootGeometry can
probably reduce the cost when a RootGeometry is not needed for any
IntersectionObservation of an observer.

Bug: 1400495
Change-Id: I295ce633b83c219861788f60498c133be4a85fab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4975516
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214953}