Backlight OCR: Implemented the `DocumentUpdated` API
In an effort to make PDFs in the Media App on ChromeOS (AKA Backlight,)
accessible, we aim to acquire images from every page of the PDF,
send them to an OCR Service and build an accessibility tree from
the results.

This patch implements the API that
informs the accessibility code that some of
the page of the PDF have been updated
and are ready for OCRing.

Design doc:
https://docs.google.com/document/d/1k8OH4XGjZrLVA705Xfy6vzSwGL9DkfRt9DG5s66lRjQ/edit?usp=sharing&resourcekey=0-K_9Y1EMhNoo8Ez9RPbQNMA

R=aleventhal@chromium.org

AX-Relnotes: n/a.
bug: b/289012145
Change-Id: I1ef753bcd7839b097c373b1879d873161108260f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4943831
Auto-Submit: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Patti Lor <patricialor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214955}