[EditContext] Fix bugs with Android touch keyboard composition
Fix two bugs with Android keyboard composition in EditContext.

The first bug is encountered after committing an autosuggested word
and then hitting backspace, when the backspace fails to delete
the character.

The cause is that the deletion calls
WebLocalFrameImpl::SetCompositionFromExistingText, which does not
call into EditContext. Fix this by implementing
EditContext::DeleteSurroundingText.

The second bug is encountered when long-pressing a word in the
middle of a sentence and then choosing an autosuggested word,
at which point the word is not replaced properly.
The cause of the bug is a misuse of wtf::String::Substring.
Fix it by passing the correctly passing the length as the
second parameter.

Bug: 1468568,999184
Change-Id: Icdeb08296f2f9a20098e0007b4bcac1e6536a373
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4919932
Reviewed-by: Anupam Snigdha <snianu@microsoft.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1215045}