Make icons in the title bar of PWAs show as disabled in inactive widgets again.
In crbug.com/1414432 while fixing the color of toolbar buttons in
PWA title bars on Mac to properly show up as disabled on macOS,
I accidentally broke the same behavior on other platforms (not
realizing that this was what previously the behavior was there). To
bring back this behavior to all platforms, this CL add a
SetAppearDisabledInInactiveWidget method to LabelButton which overrides
the value of PlatformStyle::kInactiveWidgetControlsAppearDisabled. This
is then used on all the buttons in the PWA title bar to get the
desired behavior on all platforms.

Bug: 1489460
Change-Id: I3e28546060a3f4d5cc44a35b1ce8f2d1d3612ba9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4973415
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215067}