iwa: Implement the installed IWA file structure.
Now we are copying the file that we intend to install into profile
directory. We create the 'iwa' directory in the profile dir and there
create a randomly named directory to store the signed web bundle there.

The caller is responsible to delete the original .swbn file
(if necessary).

So far it works for InstalledBundle only.

This CL introduces the concept of the .swbn ownership: is the file is
owned by Chrome, then Chrome is responsible for removing the file
on uninstallation or failed installation.

The success status of the install and prepare update commands now
include the IsolatedWebAppLocation to the place where the installed
.swbn is located.

Future work:
- remove files if pending update fails
- add garbage collection for failed deletions

Bug: b:287594073
Test: Added a unit test and executed already existing tests.
Change-Id: I1b73e1091e2b778a5473d01fd6ee46acf770680e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4637289
Reviewed-by: Christian Flach <cmfcmf@chromium.org>
Commit-Queue: Oleksandr Peletskyi <peletskyi@google.com>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Gavin Williams <gavinwill@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215122}