overview glanceable: add glanceable chip view
This cl adds the glanceable chip view component.

The glanceable chip contains an app's icon, name, content
description, and an optional call to action.

When right click or long tap the chip, there will be a removal chip
context menu.

The accessibility will be implemented in the following cls.

Test the view in the style viewer.

Bug: b:305812017
Change-Id: I47aeda208b08066f4b43ffb97c382daace023674
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4953894
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215125}